chore: prevent full page rerender on sort/filter
